Move maven tests closer to the implementation
The current `main_test.go` for fetcher is getting very long (and I'd like to refactor main.go to have a lot less logic and be easier to test). Ultimately the `main_test.go` should be removed in favor of unit tests closer to the implementation.

One of the first things we can do is move some of the newly added Maven tests out of `main_test.go` into `internal/maven`.
