+++ internal/cmd/fetcher/main_test.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"context"
-	"encoding/xml"
 	"log/slog"
 	"os"
 	"path/filepath"
@@ -17,7 +16,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/bufbuild/plugins/internal/maven"
 	"github.com/bufbuild/plugins/internal/source"
 )
 
@@ -417,261 +415,4 @@ func newTestContainer(t *testing.T, root string) appext.Container {
 	require.NoError(t, err)
 	logger := slog.New(slog.NewTextHandler(testWriter{t}, &slog.HandlerOptions{Level: slog.LevelDebug}))
 	return appext.NewContainer(nameContainer, logger)
-}
-
-func TestRegenerateMavenDepsWithPluginDeps(t *testing.T) {
-	t.Parallel()
-	tmpDir := t.TempDir()
-
-	// Setup: base-plugin has Maven deps including an additional lite runtime.
-	baseDir := filepath.Join(tmpDir, "plugins", "test", "base-plugin", "v1.0.0")
-	require.NoError(t, os.MkdirAll(baseDir, 0755))
-	baseYAML := `version: v1
-name: buf.build/test/base-plugin
-plugin_version: v1.0.0
-output_languages:
-  - java
-registry:
-  maven:
-    deps:
-      - com.google.protobuf:protobuf-java:4.33.5
-    additional_runtimes:
-      - name: lite
-        deps:
-          - com.google.protobuf:protobuf-javalite:4.33.5
-          - build.buf:protobuf-javalite:4.33.5
-        opts: [lite]
-`
-	require.NoError(t, os.WriteFile(filepath.Join(baseDir, "buf.plugin.yaml"), []byte(baseYAML), 0644))
-
-	// Setup: consumer-plugin depends on base-plugin and has its own Maven
-	// deps plus a lite runtime. The Dockerfile has a maven-deps stage.
-	consumerDir := filepath.Join(tmpDir, "plugins", "test", "consumer-plugin", "v1.0.0")
-	require.NoError(t, os.MkdirAll(consumerDir, 0755))
-	consumerYAML := `version: v1
-name: buf.build/test/consumer-plugin
-plugin_version: v1.0.0
-deps:
-  - plugin: buf.build/test/base-plugin:v1.0.0
-output_languages:
-  - kotlin
-registry:
-  maven:
-    compiler:
-      kotlin:
-        version: 1.8.22
-    deps:
-      - com.google.protobuf:protobuf-kotlin:4.33.5
-    additional_runtimes:
-      - name: lite
-        deps:
-          - com.google.protobuf:protobuf-kotlin-lite:4.33.5
-        opts: [lite]
-`
-	require.NoError(t, os.WriteFile(filepath.Join(consumerDir, "buf.plugin.yaml"), []byte(consumerYAML), 0644))
-
-	dockerfile := `# syntax=docker/dockerfile:1.19
-FROM debian:bookworm AS build
-RUN echo hello
-
-FROM scratch
-COPY --from=build /app .
-ENTRYPOINT ["/app"]
-`
-	require.NoError(t, os.WriteFile(filepath.Join(consumerDir, "Dockerfile"), []byte(dockerfile), 0644))
-
-	// Run regenerateMavenDeps on the consumer plugin
-	err := regenerateMavenDeps(createdPlugin{
-		org:        "test",
-		name:       "consumer-plugin",
-		pluginDir:  filepath.Join(tmpDir, "plugins", "test", "consumer-plugin"),
-		newVersion: "v1.0.0",
-	})
-	require.NoError(t, err)
-
-	// Verify the maven-deps stage was inserted into the Dockerfile.
-	dockerfileBytes, err := os.ReadFile(filepath.Join(consumerDir, "Dockerfile"))
-	require.NoError(t, err)
-	assert.Contains(t, string(dockerfileBytes), "FROM "+maven.MavenImage+" AS maven-deps")
-	assert.Contains(t, string(dockerfileBytes), "COPY --from=maven-deps /root/.m2/repository /maven-repository")
-
-	// Read and parse pom.xml to verify deps include versions.
-	pomBytes, err := os.ReadFile(filepath.Join(consumerDir, "pom.xml"))
-	require.NoError(t, err)
-	var pom testPOMProject
-	require.NoError(t, xml.Unmarshal(pomBytes, &pom))
-	var depVersions []string
-	for _, dep := range pom.Dependencies {
-		depVersions = append(depVersions, dep.String())
-	}
-	// Consumer's own deps should be present.
-	assert.Contains(t, depVersions, "com.google.protobuf:protobuf-kotlin:4.33.5")
-	assert.Contains(t, depVersions, "com.google.protobuf:protobuf-kotlin-lite:4.33.5")
-	// Base plugin's main deps should be merged in.
-	assert.Contains(t, depVersions, "com.google.protobuf:protobuf-java:4.33.5")
-	// Base plugin's lite runtime deps should be merged into the
-	// matching lite runtime section.
-	assert.Contains(t, depVersions, "com.google.protobuf:protobuf-javalite:4.33.5")
-	assert.Contains(t, depVersions, "build.buf:protobuf-javalite:4.33.5")
-}
-
-func TestMergeDepsMavenDepsTransitive(t *testing.T) {
-	t.Parallel()
-	tmpDir := t.TempDir()
-
-	// Setup: grandparent -> parent -> child chain
-	grandparentDir := filepath.Join(tmpDir, "plugins", "test", "grandparent", "v1.0.0")
-	require.NoError(t, os.MkdirAll(grandparentDir, 0755))
-	grandparentYAML := `version: v1
-name: buf.build/test/grandparent
-plugin_version: v1.0.0
-output_languages:
-  - java
-registry:
-  maven:
-    deps:
-      - org.example:grandparent-dep:1.0.0
-`
-	require.NoError(t, os.WriteFile(filepath.Join(grandparentDir, "buf.plugin.yaml"), []byte(grandparentYAML), 0644))
-
-	parentDir := filepath.Join(tmpDir, "plugins", "test", "parent", "v1.0.0")
-	require.NoError(t, os.MkdirAll(parentDir, 0755))
-	parentYAML := `version: v1
-name: buf.build/test/parent
-plugin_version: v1.0.0
-deps:
-  - plugin: buf.build/test/grandparent:v1.0.0
-output_languages:
-  - java
-registry:
-  maven:
-    deps:
-      - org.example:parent-dep:1.0.0
-`
-	require.NoError(t, os.WriteFile(filepath.Join(parentDir, "buf.plugin.yaml"), []byte(parentYAML), 0644))
-
-	childDir := filepath.Join(tmpDir, "plugins", "test", "child", "v1.0.0")
-	require.NoError(t, os.MkdirAll(childDir, 0755))
-	childYAML := `version: v1
-name: buf.build/test/child
-plugin_version: v1.0.0
-deps:
-  - plugin: buf.build/test/parent:v1.0.0
-output_languages:
-  - java
-registry:
-  maven:
-    deps:
-      - org.example:child-dep:1.0.0
-`
-	require.NoError(t, os.WriteFile(filepath.Join(childDir, "buf.plugin.yaml"), []byte(childYAML), 0644))
-
-	// Parse the child plugin config
-	childConfig, err := bufremotepluginconfig.ParseConfig(
-		filepath.Join(childDir, "buf.plugin.yaml"),
-	)
-	require.NoError(t, err)
-
-	// Merge transitive deps
-	pluginsDir := filepath.Join(tmpDir, "plugins")
-	err = maven.MergeTransitiveDeps(childConfig, pluginsDir)
-	require.NoError(t, err)
-
-	// Child should now have all three deps: child-dep, parent-dep,
-	// grandparent-dep (transitive through parent)
-	var artifactIDs []string
-	for _, dep := range childConfig.Registry.Maven.Deps {
-		artifactIDs = append(artifactIDs, dep.ArtifactID)
-	}
-	assert.Contains(t, artifactIDs, "child-dep")
-	assert.Contains(t, artifactIDs, "parent-dep")
-	assert.Contains(t, artifactIDs, "grandparent-dep")
-}
-
-func TestDeduplicateAllDeps(t *testing.T) {
-	t.Parallel()
-	tests := []struct {
-		name    string
-		deps    []bufremotepluginconfig.MavenDependencyConfig
-		want    []bufremotepluginconfig.MavenDependencyConfig
-		wantErr string
-	}{
-		{
-			name: "no duplicates",
-			deps: []bufremotepluginconfig.MavenDependencyConfig{
-				{GroupID: "com.example", ArtifactID: "a", Version: "1.0"},
-				{GroupID: "com.example", ArtifactID: "b", Version: "1.0"},
-			},
-			want: []bufremotepluginconfig.MavenDependencyConfig{
-				{GroupID: "com.example", ArtifactID: "a", Version: "1.0"},
-				{GroupID: "com.example", ArtifactID: "b", Version: "1.0"},
-			},
-		},
-		{
-			name: "exact duplicate removed",
-			deps: []bufremotepluginconfig.MavenDependencyConfig{
-				{GroupID: "com.example", ArtifactID: "a", Version: "1.0"},
-				{GroupID: "com.example", ArtifactID: "a", Version: "1.0"},
-			},
-			want: []bufremotepluginconfig.MavenDependencyConfig{
-				{GroupID: "com.example", ArtifactID: "a", Version: "1.0"},
-			},
-		},
-		{
-			name: "version conflict returns error",
-			deps: []bufremotepluginconfig.MavenDependencyConfig{
-				{GroupID: "com.example", ArtifactID: "a", Version: "1.0"},
-				{GroupID: "com.example", ArtifactID: "a", Version: "2.0"},
-			},
-			wantErr: "duplicate Maven dependency com.example:a with conflicting versions",
-		},
-		{
-			name: "different classifiers are distinct",
-			deps: []bufremotepluginconfig.MavenDependencyConfig{
-				{GroupID: "com.example", ArtifactID: "a", Version: "1.0", Classifier: "sources"},
-				{GroupID: "com.example", ArtifactID: "a", Version: "1.0"},
-			},
-			want: []bufremotepluginconfig.MavenDependencyConfig{
-				{GroupID: "com.example", ArtifactID: "a", Version: "1.0", Classifier: "sources"},
-				{GroupID: "com.example", ArtifactID: "a", Version: "1.0"},
-			},
-		},
-		{
-			name: "nil input",
-			deps: nil,
-			want: nil,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-			config := &bufremotepluginconfig.MavenRegistryConfig{
-				Deps: tt.deps,
-			}
-			err := maven.DeduplicateAllDeps(config)
-			if tt.wantErr != "" {
-				require.Error(t, err)
-				assert.Contains(t, err.Error(), tt.wantErr)
-				return
-			}
-			require.NoError(t, err)
-			assert.Equal(t, tt.want, config.Deps)
-		})
-	}
-}
-
-// testPOMProject mirrors the Maven POM structure for test assertions.
-type testPOMProject struct {
-	XMLName      xml.Name  `xml:"project"`
-	Dependencies []testDep `xml:"dependencies>dependency"`
-}
-
-type testDep struct {
-	GroupID    string `xml:"groupId"`
-	ArtifactID string `xml:"artifactId"`
-	Version    string `xml:"version"`
-}
-
-func (d testDep) String() string {
-	return d.GroupID + ":" + d.ArtifactID + ":" + d.Version
 }
